Point both probing algorithms to the same src code
