+++ probdiffeq/_probdiffeq/ssm_impl_blockdiag.py
@@ -1,4 +1,10 @@
-from probdiffeq._probdiffeq import problems, ssm_impl_api, taylor_points, utilities
+from probdiffeq._probdiffeq import (
+    probing,
+    problems,
+    ssm_impl_api,
+    taylor_points,
+    utilities,
+)
 from probdiffeq.backend import func, linalg, np, random, structs, tree
 from probdiffeq.backend.typing import Any, Array, Callable, Sequence, TypeVar
 from probdiffeq.util import cholesky_util
@@ -315,7 +321,7 @@ class BlockDiagLatentCondProjected(ssm_impl_api.AbstractLatentCondProjected):
 
         # Posteriors
         matvec_ensembles = func.vmap(matvec_ndmd)(ensembles)
-        C = self._blockdiag_cholesky_from_ensembles(matvec_ensembles, bias=self.bias)
+        C = probing.blockdiag_cholesky_from_ensembles(matvec_ensembles, bias=self.bias)
         noise = BlockDiagNormal(cond_mean, C, rv.tree_flatten)
 
         # Backward linear operator
@@ -327,37 +333,11 @@ class BlockDiagLatentCondProjected(ssm_impl_api.AbstractLatentCondProjected):
 
         # Marginals (redo samples for whichever reason)
         matvec_ensembles = func.vmap(self.A.matvec_ndmd)(ensembles)
-        C = self._blockdiag_cholesky_from_ensembles(matvec_ensembles, bias=self.bias)
+        C = probing.blockdiag_cholesky_from_ensembles(matvec_ensembles, bias=self.bias)
         observed = BlockDiagNormal(obs_mean, C, self.noise.tree_flatten)
 
         # Group and return
         return observed, cond
-
-    @staticmethod
-    def _blockdiag_cholesky_from_ensembles(ensembles_smd, bias: bool):
-        S, _n, _d = ensembles_smd.shape
-
-        # Center the ensembles
-        ensembles_smd -= ensembles_smd.mean(axis=0, keepdims=True)
-
-        def ensemble_to_sample_cholesky(s):
-            """Compute a sample Cholesky factor from ensembles."""
-            num, _n = s.shape
-            s = s / np.sqrt(num) if bias else s / np.sqrt(num - 1)
-            return linalg.qr_r(s).T
-
-        # The QR decomposition is why we assume S >= n,
-        # so let's check it briefly:
-        _S, m, _d = ensembles_smd.shape
-        if m > S:
-            msg = "The function requires at least as many ensembles as Taylor coefficients."
-            msg += f" Received: S={S} < m={m}, which violates this assumption."
-            raise ValueError(msg)
-
-        # Assume ensembles are shape (S, n, d), so we batch along d
-        # but since we also want the output to be (d, n, n), the out_axes is 0.
-        transform = func.vmap(ensemble_to_sample_cholesky, in_axes=-1, out_axes=0)
-        return transform(ensembles_smd)
 
     def merge(self, other, /):
         raise NotImplementedError
